Apply parallelism (#17)
Adds parallelism using the concurrent.futures module. It doesn't break the existing API across previous versions. It does change the default behaviour though to parallelise operations across the available number of processors on the system it is being run. This could present a speedup from what the default behaviour used to be - which simply used a for-loop to perform operations linearly across individuals in a generation.